+++ ecs-agent/api/container/status/containerstatus_test.go
@@ -22,11 +22,17 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestShouldReportToBackend(t *testing.T) {
 	// ContainerStatusNone is not reported to backend
 	var containerStatus ContainerStatus
+	assert.False(t, containerStatus.ShouldReportToBackend(ContainerRunning))
+	assert.False(t, containerStatus.ShouldReportToBackend(ContainerResourcesProvisioned))
+
+	// ContainerManifestPulled is not reported to backend
+	containerStatus = ContainerManifestPulled
 	assert.False(t, containerStatus.ShouldReportToBackend(ContainerRunning))
 	assert.False(t, containerStatus.ShouldReportToBackend(ContainerResourcesProvisioned))
 
@@ -64,6 +70,11 @@ func TestBackendStatus(t *testing.T) {
 	assert.Equal(t, containerStatus.BackendStatus(ContainerRunning), ContainerStatusNone)
 	assert.Equal(t, containerStatus.BackendStatus(ContainerResourcesProvisioned), ContainerStatusNone)
 
+	// BackendStatus is still ContainerStatusNone when container status is ContainerManifestPulled
+	containerStatus = ContainerManifestPulled
+	assert.Equal(t, containerStatus.BackendStatus(ContainerRunning), ContainerStatusNone)
+	assert.Equal(t, containerStatus.BackendStatus(ContainerResourcesProvisioned), ContainerStatusNone)
+
 	// BackendStatus is still ContainerStatusNone when container status is ContainerPulled
 	containerStatus = ContainerPulled
 	assert.Equal(t, containerStatus.BackendStatus(ContainerRunning), ContainerStatusNone)
@@ -97,19 +108,55 @@ type testContainerStatus struct {
 	SomeStatus ContainerStatus `json:"status"`
 }
 
-func TestUnmarshalContainerStatus(t *testing.T) {
-	status := ContainerStatusNone
-
-	err := json.Unmarshal([]byte(`"RUNNING"`), &status)
-	if err != nil {
-		t.Error(err)
+// Tests that pointers to a struct that contains ContainerStatus are marshaled correctly.
+func TestMarshalContainerStatusNestedPointer(t *testing.T) {
+	tcs := []struct {
+		status   ContainerStatus
+		expected string
+	}{
+		{status: ContainerStatusNone, expected: "NONE"},
+		{status: ContainerManifestPulled, expected: "MANIFEST_PULLED"},
+		{status: ContainerPulled, expected: "PULLED"},
+		{status: ContainerCreated, expected: "CREATED"},
+		{status: ContainerRunning, expected: "RUNNING"},
+		{status: ContainerResourcesProvisioned, expected: "RESOURCES_PROVISIONED"},
+		{status: ContainerStopped, expected: "STOPPED"},
 	}
-	if status != ContainerRunning {
-		t.Error("RUNNING should unmarshal to RUNNING, not " + status.String())
+	for _, tc := range tcs {
+		marshaled, err := json.Marshal(&testContainerStatus{SomeStatus: tc.status})
+		require.NoError(t, err)
+		assert.Equal(t, fmt.Sprintf(`{"status":%q}`, tc.status), string(marshaled))
 	}
+}
 
+// Tests that ContainerStatus is unmarshaled correctly.
+func TestUnmarshalContainerStatusSimple(t *testing.T) {
+	tcs := []struct {
+		statusString string
+		expected     ContainerStatus
+	}{
+		{statusString: "NONE", expected: ContainerStatusNone},
+		{statusString: "MANIFEST_PULLED", expected: ContainerManifestPulled},
+		{statusString: "PULLED", expected: ContainerPulled},
+		{statusString: "CREATED", expected: ContainerCreated},
+		{statusString: "RUNNING", expected: ContainerRunning},
+		{statusString: "RESOURCES_PROVISIONED", expected: ContainerResourcesProvisioned},
+		{statusString: "STOPPED", expected: ContainerStopped},
+	}
+	for _, tc := range tcs {
+		t.Run(tc.statusString, func(t *testing.T) {
+			status := ContainerStatusNone
+			err := json.Unmarshal([]byte(fmt.Sprintf("%q", tc.statusString)), &status)
+			require.NoError(t, err)
+			assert.Equal(t, tc.expected, status)
+		})
+	}
+}
+
+// Tests that structs that have a ContainerStatus field are unmarshaled correctly.
+func TestUnmarshalContainerStatusNested(t *testing.T) {
 	var test testContainerStatus
-	err = json.Unmarshal([]byte(`{"status":"STOPPED"}`), &test)
+	err := json.Unmarshal([]byte(`{"status":"STOPPED"}`), &test)
 	if err != nil {
 		t.Error(err)
 	}
@@ -168,6 +215,110 @@ func TestUnmarshalContainerHealthStatus(t *testing.T) {
 			err := json.Unmarshal([]byte(tc.String), &status)
 			assert.NoError(t, err)
 			assert.Equal(t, status, tc.Status)
+		})
+	}
+}
+
+// Tests that all container statuses are marshaled to JSON with a quoted string.
+// Also tests that JSON marshaled container status can be unmarshaled.
+func TestContainerStatusMarshalUnmarshalJSON(t *testing.T) {
+	for strStatus, status := range containerStatusMap {
+		t.Run(fmt.Sprintf("marshal-unmarshal %v", strStatus), func(t *testing.T) {
+			marshaled, err := json.Marshal(status)
+			require.NoError(t, err)
+			require.Equal(t, fmt.Sprintf("%q", strStatus), string(marshaled))
+
+			var unmarshaled ContainerStatus
+			err = json.Unmarshal(marshaled, &unmarshaled)
+			require.NoError(t, err)
+			require.Equal(t, status, unmarshaled)
+		})
+	}
+}
+
+// Tests that a container status marshaled as text can be unmarshaled.
+func TestContainerStatusMarshalUnmarshalText(t *testing.T) {
+	for strStatus, status := range containerStatusMap {
+		t.Run(fmt.Sprintf("marshal-unmarshal %v", strStatus), func(t *testing.T) {
+			marshaled, err := status.MarshalText()
+			require.NoError(t, err)
+			require.Equal(t, fmt.Sprintf("%s", strStatus), string(marshaled))
+
+			var unmarshaled ContainerStatus
+			err = unmarshaled.UnmarshalText(marshaled)
+			require.NoError(t, err)
+			require.Equal(t, status, unmarshaled)
+		})
+	}
+}
+
+// Tests that MarshalText works as expected for container status pointers.
+func TestContainerStatusMarshalPointer(t *testing.T) {
+	status := ContainerPulled
+	ptr := &status
+	marshaled, err := ptr.MarshalText()
+	require.NoError(t, err)
+	assert.Equal(t, "PULLED", string(marshaled))
+}
+
+// Tests that unmarshaling an invalid text to container status fails.
+func TestContainerStatusTextUnmarshalError(t *testing.T) {
+	var status ContainerStatus
+	assert.EqualError(t, status.UnmarshalText([]byte("invalidStatus")),
+		"container status text unmarshal: unrecognized status: invalidStatus")
+}
+
+// Tests that string based statuses are used when a map with container status as keys is
+// marshaled to JSON.
+func TestContainerStatusKeyMarshal(t *testing.T) {
+	someMap := map[ContainerStatus]string{
+		ContainerStatusNone:           "",
+		ContainerPulled:               "",
+		ContainerCreated:              "",
+		ContainerRunning:              "",
+		ContainerResourcesProvisioned: "",
+		ContainerStopped:              "",
+	}
+	marshaled, err := json.Marshal(someMap)
+	require.NoError(t, err)
+
+	var unmarshaledMap map[string]string
+	err = json.Unmarshal(marshaled, &unmarshaledMap)
+	require.NoError(t, err)
+
+	assert.Equal(t, map[string]string{
+		`NONE`:                  "",
+		`PULLED`:                "",
+		`CREATED`:               "",
+		`RUNNING`:               "",
+		`RESOURCES_PROVISIONED`: "",
+		`STOPPED`:               "",
+	}, unmarshaledMap)
+
+	var unmarshaled map[ContainerStatus]string
+	err = json.Unmarshal(marshaled, &unmarshaled)
+	require.NoError(t, err)
+	assert.Equal(t, someMap, unmarshaled)
+}
+
+// Tests that JSON unmarshal of container status is backwards-compatible with legacy integer
+// based representations for JSON object keys.
+func TestContainerStatusJSONUnmarshalInt(t *testing.T) {
+	tcs := map[string]ContainerStatus{
+		`"0"`: ContainerStatusNone,
+		`"1"`: ContainerPulled,
+		`"2"`: ContainerCreated,
+		`"3"`: ContainerRunning,
+		`"4"`: ContainerResourcesProvisioned,
+		`"5"`: ContainerStopped,
+		`"6"`: ContainerZombie,
+	}
+	for intStatus, status := range tcs {
+		t.Run(fmt.Sprintf("%s - %s", intStatus, status.String()), func(t *testing.T) {
+			var unmarshaled ContainerStatus
+			err := json.Unmarshal([]byte(intStatus), &unmarshaled)
+			require.NoError(t, err)
+			assert.Equal(t, status, unmarshaled)
 		})
 	}
 }