+++ ecs-init/docker/docker.go
@@ -59,7 +59,7 @@ const (
 	minBackoffDuration = time.Second
 	// maxBackoffDuration specifies the maximum backoff duration for ping to
 	// return a success response from docker socket
-	maxBackoffDuration = 5 * time.Second
+	maxBackoffDuration = 3 * time.Second
 	// backoffJitterMultiple specifies the backoff jitter multiplier
 	// coefficient when pinging the docker socket
 	backoffJitterMultiple = 0.2
@@ -290,6 +290,7 @@ func (c *client) getContainerConfig(envVarsFromFiles map[string]string) *godocke
 		"ECS_AVAILABLE_LOGGING_DRIVERS":         `["json-file","syslog","awslogs","fluentd","none"]`,
 		"ECS_ENABLE_TASK_IAM_ROLE":              "true",
 		"ECS_ENABLE_TASK_IAM_ROLE_NETWORK_HOST": "true",
+		"ECS_EBSTA_SUPPORTED":                   "true",
 		"ECS_AGENT_LABELS":                      "",
 		"ECS_VOLUME_PLUGIN_CAPABILITIES":        `["efsAuth"]`,
 	}
@@ -325,6 +326,17 @@ func (c *client) getContainerConfig(envVarsFromFiles map[string]string) *godocke
 	if config.RunningInExternal() {
 		// Task networking is not supported when not running on EC2. Explicitly disable since it's enabled by default.
 		envVariables["ECS_ENABLE_TASK_ENI"] = "false"
+	}
+	if dockerVersionCompare(dockerAPIVersion, enableTaskENIDockerClientAPIVersion) < 0 {
+		// Task networking is not supported before docker API version 1.25
+		envVariables["ECS_ENABLE_TASK_ENI"] = "false"
+	}
+
+	if !isPathValid(config.MountDirectoryEBS(), true) {
+		// EBS Task Attach (EBSTA) is not supported for external instances
+		// If EBS mount directory fails to get created, tasks requiring EBSTA can not be supported
+		// Disable EBSTA Support for these cases
+		envVariables["ECS_EBSTA_SUPPORTED"] = "false"
 	}
 
 	var env []string