+++ scripts/ecs-anywhere-install.sh
@@ -637,7 +637,7 @@ find-copy-certs-exec() {
 }
 
 download-ssm-binaries-exec() {
-    BINARY_VERSION="3.2.1630.0"
+    BINARY_VERSION="3.2.2303.0"
     BINARY_PATH="/var/lib/ecs/deps/execute-command/bin/${BINARY_VERSION}"
     BINARY_DOWNLOAD_PATH="ssm-binaries"